Change data filepath
To make sure your IDE is able to find the data in the assignment 6 folder, change the root that the IDE is working in (for Spyder in the top right corner). Don't change the filepath itself as we will all have to change it back to work with the program again.